# replace legacy string with lolly_string