--- a/Kernel/Abstractions/basic.cpp
+++ b/Kernel/Abstractions/basic.cpp
@@ -12,6 +12,7 @@
 #include "basic.hpp"
 #include "analyze.hpp"
 #include "string.hpp"
+#include <string.h>
 
 int
 hash (pointer ptr) {
@@ -25,13 +26,14 @@ string the_report;
 
 void
 tm_throw (const char* msg) {
-  the_exception= msg;
+  string msg_  = as_string (msg);
+  the_exception= msg_;
   // the_report   = get_crash_report (msg);
   cout << "Throwing " << msg << LF;
   cout << "-------------------------------------------------\n";
   cout << the_report << LF;
   cout << "-------------------------------------------------\n";
-  throw string (msg);
+  throw string (msg_);
 }
 
 void