--- a/Kernel/Types/string.cpp
+++ b/Kernel/Types/string.cpp
@@ -156,6 +156,10 @@ as_string (unsigned long int i) {
 
 string
 as_string (double x) {
+#if defined(__clang__) && (__clang_major__ < 14)
+  std::string str= std::to_string (x);
+  return string (string_view (str.size (), str.data ()));
+#else
   char buffer[64];
   auto [ptr, ec]= std::to_chars (buffer, buffer + 32, x);
   if (ec == std::errc ()) {
@@ -164,6 +168,7 @@ as_string (double x) {
   else {
     return string ();
   }
+#endif
 }
 
 string