--- a/System/Classes/file_url.cpp
+++ b/System/Classes/file_url.cpp
@@ -16,7 +16,7 @@ file_url::file_url (string name)
     : url (url_root ("file") * url_get_name (name, URL_SYSTEM)) {}
 
 file_url::file_url (const char* name)
-    : url (url_root ("file") * url_get_name (string (name), URL_SYSTEM)) {}
+    : url (url_root ("file") * url_get_name (as_string (name), URL_SYSTEM)) {}
 
 string
 file_url::concretize () {