--- a/System/Classes/url.cpp
+++ b/System/Classes/url.cpp
@@ -44,21 +44,24 @@ static inline bool
 is_tuple (url_tree t, string s) {
   return (t->op == URL_TUPLE) && (N (t) >= 1) && (t[0] == s);
 }
+template <size_t Ns>
 static inline bool
-is_tuple (url_tree t, const char* s) {
+is_tuple (url_tree t, const char (&s)[Ns]) {
   return (t->op == URL_TUPLE) && (N (t) >= 1) && (t[0] == s);
 }
 static inline bool
 is_tuple (url_tree t, string s, int n) {
   return (t->op == URL_TUPLE) && (N (t) == (n + 1)) && (t[0] == s);
 }
+template <size_t Ns>
 static inline bool
-is_tuple (url_tree t, const char* s, int n) {
+is_tuple (url_tree t, const char (&s)[Ns], int n) {
   return (t->op == URL_TUPLE) && (N (t) == (n + 1)) && (t[0] == s);
 }
 
 url::url () : rep (tm_new<url_rep> (url_tuple ("none"))) {}
-url::url (const char* name) : rep (tm_new<url_rep> (url_system (name)->t)) {}
+url::url (const char* name)
+    : rep (tm_new<url_rep> (url_system (as_string (name))->t)) {}
 url::url (string name) : rep (tm_new<url_rep> (url_system (name)->t)) {}
 url::url (string path_name, string name)
     : rep (tm_new<url_rep> (url_system (path_name, name)->t)) {}
@@ -326,7 +329,7 @@ url
 url_pwd () {
   char path[TB_PATH_MAXN];
   if (tb_directory_current (path, TB_PATH_MAXN)) {
-    return url_system (path);
+    return url_system (as_string(path));
   }
   else {
     TM_FAILED ("FAILED to get pwd");