--- a/System/IO/tm_ostream.hpp
+++ b/System/IO/tm_ostream.hpp
@@ -15,7 +15,6 @@
 // #include "url.hpp"
 #include "string.hpp"
 #include <cstdio>
-#include <cstdlib>
 class tm_ostream;
 
 class tm_ostream_rep {