--- a/System/Memory/impl/mi_malloc.cpp
+++ b/System/Memory/impl/mi_malloc.cpp
@@ -10,8 +10,8 @@
  ******************************************************************************/
 
 #include "assert.h"
-#include "basic.hpp"
 #include "fast_alloc.hpp"
+#include "tm_ostream.hpp"
 #include <errno.h>
 #include <mimalloc.h>