--- a/lolly/data/numeral.hpp
+++ b/lolly/data/numeral.hpp
@@ -12,6 +12,7 @@
 #pragma once
 
 #include "string.hpp"
+#include "basic.hpp"
 #include <stdint.h>
 
 namespace lolly {