… Kernel/Abstractions/basic.hpp
… Kernel/Abstractions/classdef.hpp
… System/Misc/sys_utils.cpp
… bench/Kernel/Types/string_bench.cpp
… lolly/data/json.hpp
… lolly/data/lolly_string.hpp
… lolly/data/lolly_string.ipp
… lolly/data/lolly_tree.hpp
… lolly/data/numeral.cpp
… lolly/hash/uuid.cpp
… lolly/io/http.cpp
… lolly/system/timer.hpp
… tests/Kernel/Abstractions/blackbox_test.cpp
… tests/System/Classes/tm_timer_test.cpp
… tests/System/Classes/url_test.cpp