## Changes

* reform `basic.hpp` thus `tm_ostream.hpp` and `string.hpp` can include part of its content
* remove `string(const char*)` constructor. Please use `as_string` instead, except constructing from string literal.
* re-implement `as_string`
* some minor fix in `lolly_string` and test of blackbox

## Performance

### Before

|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|              112.97 |        8,851,657.81 |    1.2% |      0.27 | `string from int_16`
|              113.53 |        8,808,509.70 |    0.5% |      0.27 | `string from int_32`
|              116.08 |        8,615,090.92 |    0.5% |      0.28 | `string from int_64`
|              224.31 |        4,458,183.60 |    1.3% |      0.54 | `string from double`

### After
|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|              107.95 |        9,263,974.89 |    1.0% |      0.26 | `string from int_16`
|              105.45 |        9,483,062.34 |    1.4% |      0.25 | `string from int_32`
|              105.85 |        9,447,547.71 |    1.2% |      0.25 | `string from int_64`
|              162.49 |        6,154,251.13 |    1.7% |      0.39 | `string from double`