@darcy-shen 2024-03-18
> Failure of lint ci may be erroneous, such indentation cannot pass formatter installed on my computer

It seems to be a clang-format bug.