@darcy-shen 2024-03-18
Replacing the legacy string with lolly_string should be done in a project, for example project 21.

It should be done in several small pull requests.