@jingkaimori 2024-03-18
> Which pr should I merge first to minimize the code changes (git conflicts)?

#295 , because the two pr is orthodox, thus small pr comes first can reduce changes