@darcy-shen 2024-03-18
> build and test ok on my macOS 14, what C++ standard are you using?

Here is the clues to fix ci on macOS
https://github.com/llvm/llvm-project/issues/59374