@darcy-shen 2024-03-18
You can split `basic.hpp` in `21_1`, and then re-implement as_string in `21_2`